멀티스레딩 로딩 추가 프리팹 SceneOutliner 생성 추가 UI Tool 오류 수정(복사 생성자 잘못되어 있어, 일부만 추가되는 형태였음) 필터 정리 TreeUI 외부 Drag/Drop 기능 추가 로그 기능 추가 TimeMeasure를 통한 실행시간 측정 기능 추가

+++ Project/Script/Deux_Attack.cpp
@@ -8,15 +8,11 @@
 #include <Engine/CAnimator3D.h>
 
 Deux_Attack::Deux_Attack()
-	: CState{ L"ATTACK" }
+	: CState{L"ATTACK"}
 	, m_pAnimation(nullptr)
-	, m_pOwnerMGR(nullptr)
-{
-}
+	, m_pOwnerMGR(nullptr) {}
 
-Deux_Attack::~Deux_Attack()
-{
-}
+Deux_Attack::~Deux_Attack() {}
 
 void Deux_Attack::Enter()
 {
@@ -33,7 +29,6 @@ void Deux_Attack::Enter()
 	}
 	m_pAnimation->SetLerpTime(1.5f);
 	GetOwner()->Animator3D()->Play(m_pAnimation->GetName(), false);
-
 }
 
 void Deux_Attack::Update()
@@ -53,7 +48,7 @@ void Deux_Attack::Update()
 	}
 	else
 	{
-		float fAnimationLength = m_pAnimation->GetEndTime() - m_pAnimation->GetStartTime();
+		float fAnimationLength = (float)(m_pAnimation->GetEndTime() - m_pAnimation->GetStartTime());
 		if (fAnimationLength <= CState::GetTimer())
 		{
 			GetOwner()->GetScript<FieldMonsteScript>()->SetCurAnimationDone();
